requirements.txt
google-cloud-bigquery==1.9.0
Flask==1.0.2
numpy==1.16.4
pandas==0.24.2
pytest==4.6.3
